obs/requirements.txt
obs-websocket-py
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
asyncio-mqtt
